backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python-headless==4.8.1.78
pandas==2.1.3
openpyxl==3.1.2
pillow==10.1.0
aiosqlite==0.19.0
google-generativeai==0.3.2
asyncpg==0.29.0
python-docx==0.8.11
PyMuPDF==1.23.8